// CRMHotelFront/src/app/interfaces/usuario.interfaces.ts
export interface Usuario {
    id: number;
    nombre: string;
    apellidos: string;
    fecha_nacimiento: Date;
    dni: string;
    email: string;
    username: string;
    password: string;
    rol: string;
    sexo: string;
    telefono: number;
    pais: string;
    direccion: string;
    ciudad: string;
    cod_postal: string;
}